services/analytics/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
paho-mqtt
redis
psycopg2-binary
python-dotenv
